// fe/src/lib/api/generated/model/exceptionCreateRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { AvailabilityModeScope } from './availabilityModeScope';

export interface ExceptionCreateRequest {
  ends_at: string;
  mode_scope: AvailabilityModeScope;
  reason?: string;
  starts_at: string;
}